adding test

+++ src/main/java/bankaccount/mybank/MainApp.java
@@ -1,5 +1,7 @@
 package bankaccount.mybank;
 
+import java.math.BigDecimal;
+
 /**
  * Hello world!
  *
@@ -11,10 +13,10 @@ public class MainApp
         Account account = new Account();
 
         //Add 100 to Account
-        account.deposit(Double.valueOf(100));
+        account.deposit(BigDecimal.valueOf(100));
         
         //withdraw 50 from account
-        account.withdraw(Double.valueOf(50));
+        account.withdraw(BigDecimal.valueOf(50));
         
         //print operations history
         System.out.println( "------------------------------------------------");
